Define many content layout types for FindServerScreen
