+++ features/auth/ui/src/main/kotlin/com/nasdroid/auth/ui/Navigation.kt
@@ -47,6 +47,7 @@ fun NavGraphBuilder.authNavigation(
                 },
                 modifier = modifier,
                 contentPadding = contentPadding,
+                windowSizeClass = windowSizeClass,
             )
         }
         composable(
