+++ features/auth/ui/src/main/kotlin/com/nasdroid/auth/ui/register/find/FindServerScreen.kt
@@ -4,47 +4,28 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Dns
-import androidx.compose.material3.Button
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.material3.dynamicDarkColorScheme
 import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
+import androidx.compose.material3.windowsizeclass.WindowHeightSizeClass
 import androidx.compose.material3.windowsizeclass.WindowSizeClass
+import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.tooling.preview.PreviewScreenSizes
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import com.nasdroid.auth.logic.Server
-import com.nasdroid.auth.ui.R
-import com.nasdroid.auth.ui.serverselect.SelectServerContent
 import com.nasdroid.design.MaterialThemeExt
 import org.koin.androidx.compose.koinViewModel
 
@@ -54,13 +35,16 @@ import org.koin.androidx.compose.koinViewModel
 @Composable
 fun FindServerScreen(
     onServerFound: (address: String) -> Unit,
+    windowSizeClass: WindowSizeClass,
     modifier: Modifier = Modifier,
     contentPadding: PaddingValues = PaddingValues(),
     viewModel: FindServerViewModel = koinViewModel()
 ) {
-    FindServerByAddress(
-        onServerAddressChange = onServerFound,
-        modifier = modifier.padding(contentPadding)
+    FindServerContent(
+        onServerFound = onServerFound,
+        windowSizeClass = windowSizeClass,
+        modifier = modifier,
+        contentPadding = contentPadding
     )
 }
 
@@ -71,77 +55,63 @@ fun FindServerContent(
     modifier: Modifier = Modifier,
     contentPadding: PaddingValues = PaddingValues(),
 ) {
-    FindServerByAddress(
-        onServerAddressChange = onServerFound,
-        modifier = modifier.padding(contentPadding)
-    )
-}
-
-@Composable
-internal fun FindServerByAddress(
-    onServerAddressChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val (address, onAddressChange) = rememberSaveable { mutableStateOf("") }
-    val canConnect by remember(address) { derivedStateOf { address.isNotBlank() }}
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.Center
-    ) {
-        ServerAddressField(
-            serverAddress = address,
-            onServerAddressChange = onAddressChange,
-            onKeyboardAction = {
-                if (canConnect) onServerAddressChange(address)
-            },
-            modifier = Modifier
-                .widthIn(max = 480.dp)
-                .fillMaxWidth()
-        )
-        Spacer(Modifier.height(16.dp))
-        Button(
-            onClick = { onServerAddressChange(address) },
-            modifier = Modifier
-                .widthIn(max = 480.dp)
-                .fillMaxWidth(),
-            enabled = canConnect
-        ) {
-            Text(stringResource(R.string.connect_server))
+    when {
+        windowSizeClass.widthSizeClass > WindowWidthSizeClass.Compact &&
+                windowSizeClass.heightSizeClass == WindowHeightSizeClass.Compact -> {
+            FindServerHorizontalContent(
+                onServerAddressChange = onServerFound,
+                modifier = modifier.padding(contentPadding)
+            )
+        }
+        windowSizeClass.widthSizeClass == WindowWidthSizeClass.Expanded &&
+                windowSizeClass.heightSizeClass == WindowHeightSizeClass.Expanded -> {
+            FindServerCenteredContent(
+                onServerAddressChange = onServerFound,
+                modifier = modifier.padding(contentPadding)
+            )
+        }
+        else -> {
+            FindServerVerticalContent(
+                onServerAddressChange = onServerFound,
+                modifier = modifier.padding(contentPadding)
+            )
         }
     }
 }
 
 @Composable
-internal fun ServerAddressField(
-    serverAddress: String,
-    onServerAddressChange: (String) -> Unit,
-    onKeyboardAction: () -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    error: Boolean = false
+fun FindServerVerticalContent(
+    onServerAddressChange: (address: String) -> Unit,
+    modifier: Modifier = Modifier
 ) {
-    TextField(
-        value = serverAddress,
-        onValueChange = onServerAddressChange,
-        label = { Text(stringResource(R.string.server_label)) },
-        leadingIcon = { Icon(Icons.Default.Dns, contentDescription = null) },
-        keyboardOptions = KeyboardOptions(
-            keyboardType = KeyboardType.Uri,
-            autoCorrect = false,
-            imeAction = ImeAction.Done
-        ),
-        keyboardActions = KeyboardActions { onKeyboardAction() },
-        singleLine = true,
-        enabled = enabled,
-        isError = error,
-        supportingText = if (error) {{
-            Text(stringResource(R.string.invalid_server_address))
-        }} else {
-            null
-        },
-        modifier = modifier
-    )
+    Column(modifier) {
+        FindServerByAddress(
+            onServerAddressChange = onServerAddressChange
+        )
+    }
+}
+
+@Composable
+fun FindServerHorizontalContent(
+    onServerAddressChange: (address: String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(modifier) {
+        FindServerByAddress(onServerAddressChange = onServerAddressChange)
+    }
+}
+
+@Composable
+fun FindServerCenteredContent(
+    onServerAddressChange: (address: String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Box(modifier) {
+        FindServerByAddress(
+            onServerAddressChange = onServerAddressChange,
+            modifier = Modifier.align(Alignment.Center)
+        )
+    }
 }
 
 @OptIn(ExperimentalMaterial3WindowSizeClassApi::class)
@@ -169,7 +139,7 @@ fun FindServerScreenPreview() {
             modifier = Modifier
                 .fillMaxSize()
                 .background(MaterialThemeExt.colorScheme.background),
-            contentPadding = PaddingValues(16.dp)
+            contentPadding = PaddingValues(MaterialThemeExt.paddings.large)
         )
     }
 }
